requirements.txt
# ── Core deep learning ────────────────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0

# ── Segmentation model ────────────────────────────────────────────────────────
segmentation-models-pytorch>=0.3.3

# ── Data / augmentation ───────────────────────────────────────────────────────
albumentations>=1.3.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0

# ── Visualisation ─────────────────────────────────────────────────────────────
matplotlib>=3.7.0

# ── Config ────────────────────────────────────────────────────────────────────
PyYAML>=6.0

# ── Logging ───────────────────────────────────────────────────────────────────
tensorboard>=2.13.0

# ── Optional: nicer progress bars ─────────────────────────────────────────────
rich>=13.0.0

# ── Optional: Weights & Biases experiment tracking ────────────────────────────
# wandb>=0.15.0

# ── Dev tools ─────────────────────────────────────────────────────────────────
pytest>=7.0.0
jupyter>=1.0.0
tqdm>=4.65.0